backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
click==8.1.8
distlib==0.3.9
Django==5.2
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
filelock==3.18.0
flake8==7.2.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rest-framework-simplejwt==0.0.2
rpds-py==0.24.0
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.7.4.3
uritemplate==4.1.1
virtualenv==20.30.0
django-jazzmin>=2.6.0
django-filter